to avoid circular imports, move the MixIn into a base module
